refactor: 전체 패키지 구조 리팩터링 적용
