+++ README.md
@@ -54,7 +54,7 @@
     - [ ] 2 번째 분배부터는 카드 수령 여부를 입력 받아 임의로 분배 받는다.
       - [ ] 딜러는 모든 플레이어가 추가 카드 수령을 완료 한 후 카드 1장을 분배 받는다.
         - [ ] 첫 분배에 받은 2장의 카드가 16이하이면 무조건 추가로 1장을 분배 받는다.
-          - [ ] 딜러는 추가 1장의 카드를 분배 받은 이후, 추가로 분배를 받지 아니한다.
+          - [ ] 딜러는 추가 1장의 카드를 분배 받은 이후, 핸드가 17이상이 될 때까지 카드를 받는다.
 
 2. [ ] 모든 참가자들의 카드 분배가 완료 된 후 결과를 계산한다.
     - [ ] 개별 참가자들이 소유하고 있는 카드의 총합을 계산한다.
