--- src/main/java/model/Player.java
@@ -1,13 +1,0 @@
-package model;
-
-public class Player extends Participant {
-
-    public Player(Name name) {
-        super(name);
-    }
-
-    @Override
-    public boolean canHit() {
-        return !cardDeck.isBust();
-    }
-}
