+++ src/main/java/model/card/Emblem.java
@@ -1,0 +1,18 @@
+package model.card;
+
+public enum Emblem {
+    SPADE("스페이드"),
+    DIAMOND("다이아몬드"),
+    CLUB("클로버"),
+    HEART("하트");
+
+    private final String value;
+
+    Emblem(String value) {
+        this.value = value;
+    }
+
+    public String getValue() {
+        return value;
+    }
+}
