+++ src/main/java/model/participant/Player.java
@@ -1,0 +1,13 @@
+package model.participant;
+
+public class Player extends Participant {
+
+    public Player(Name name) {
+        super(name);
+    }
+
+    @Override
+    public boolean canHit() {
+        return !cardDeck.isBust();
+    }
+}
