+++ src/main/java/view/OutputView.java
@@ -1,0 +1,16 @@
+package view;
+
+public class OutputView {
+
+    private OutputView() {
+    }
+
+    public static void printNewLine() {
+        System.out.print(System.lineSeparator());
+    }
+
+    public static void print(String toBePrint) {
+        System.out.println(toBePrint);
+    }
+
+}
