--- src/test/java/model/FixedCardShuffleMachine.java
@@ -1,12 +1,0 @@
-package model;
-
-import java.util.LinkedList;
-import java.util.Queue;
-
-public class FixedCardShuffleMachine implements CardShuffleMachine{
-    @Override
-    public Queue<Card> shuffleCardDeck() {
-        return new LinkedList<>(generateAllCards());
-    }
-
-}
